+++ src/components/three/Light.tsx
