+++ src/components/three/Loader.tsx
@@ -1,6 +1,6 @@
 import { Html, useProgress } from '@react-three/drei';
 import React from 'react';
-import ProgressBar from '../ProgressBar';
+import ProgressBar from '../buttons/ProgressBar';
 
 const Loader = () => {
   const { progress } = useProgress();
