+++ src/components/three/PortfolioThree/Minimap.tsx
