+++ src/components/three/PortfolioThree/proxy.ts
