+++ src/components/three/Text.tsx
