+++ src/views/Idea/Idea.tsx
