+++ src/views/Resume/ResumeIntroduce.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { makeStyles, Grid, Typography } from '@material-ui/core';
-import ko from '../../../util/textData';
+import ko from '../../util/textData';
 
 const useStyle = makeStyles((theme) => ({
   title: {
@@ -8,32 +8,32 @@ const useStyle = makeStyles((theme) => ({
     margin: '20px 0px 10px 0px',
     color: '#104869',
   },
-  contentContainer: {
+  subContentContainer: {
     margin: '5px 0px',
   },
 }));
 
-const ResumeOpenSource = () => {
+const ResumeIntroduce = () => {
   const classes = useStyle();
   const data = ko;
 
   return (
-    <Grid item md={6} xs={12}>
+    <Grid item xs={12}>
       <Typography variant='h4' className={classes.title}>
-        Open Source.
+        Introduce.
       </Typography>
-      <div className={classes.contentContainer}>
-        <Typography>
-          <a href={data.github}>{data.github}</a>
-        </Typography>
-      </div>
-      <div className={classes.contentContainer}>
-        <Typography>
-          <a href={data.npm}>{data.npm}</a>
+      <div className={classes.subContentContainer}>
+        <Typography variant='subtitle1'>
+          {data.introduce.split('\n').map((line, idx) => (
+            <span key={idx}>
+              {line}
+              <br />
+            </span>
+          ))}
         </Typography>
       </div>
     </Grid>
   );
 };
 
-export default ResumeOpenSource;
+export default ResumeIntroduce;
