+++ src/views/Resume/ResumeWorkExperience.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { makeStyles, Grid, Typography, Divider } from '@material-ui/core';
 import { ArrowRight } from '@material-ui/icons';
-import ko from '../../../util/textData';
+import ko from '../../util/textData';
 
 const useStyle = makeStyles((theme) => ({
   title: {
